requirements.txt
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-channels==0.7.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
exceptiongroup==1.1.3
flake8==6.1.0
h11==0.14.0
httptools==0.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
zope.interface==6.0
